Renamed the directory /storages to /storage
We have a .gitignore that is crashing the name /storages so any change
done on that directory is not take into account by git add.

To avoid issue I've changed the directory name, this is the rule:

st[A-Za-z0-9][A-Za-z0-9][A-Za-z0-9][A-Za-z0-9][A-Za-z0-9][A-Za-z0-9]
