+++ SCsub
@@ -8,7 +8,7 @@ Import("env_modules")
 ecs = env_modules.Clone()
 
 # Module source files
-ecs.add_source_files(env.modules_sources, "storages/*.cpp")
+ecs.add_source_files(env.modules_sources, "memory/*.cpp")
 ecs.add_source_files(env.modules_sources, "components/*.cpp")
 ecs.add_source_files(env.modules_sources, "iterators/*.cpp")
 ecs.add_source_files(env.modules_sources, "systems/*.cpp")
